# openapi.yaml
openapi: 3.0.3
info:
  title: FHNW Connect REST API
  version: 1.0.0
  description: |
    OpenAPI specification for FHNW Connect.

    This API is derived from the uploaded README and ERD/domain model.
    It covers the core business capabilities:
    - browse and filter activities
    - view activity details
    - join and leave activities
    - favorite clubs and activities
    - manage community posts
    - create club announcements
    - inspect participation numbers

    Notes:
    - Authentication is modeled with JWT bearer tokens.
    - Role/ownership checks (student vs. club manager/admin) are left to backend authorization logic.
    - "Announcements" are modeled as community posts with category `announcement` and a `club_id`.
servers:
  - url: http://localhost:8000/api
    description: Local development
  - url: https://your-render-service.onrender.com/api
    description: Example Render deployment
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in a user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /activities:
    get:
      tags: [Activities]
      summary: List activities
      description: Supports searching and filtering by category, date, club, location, and favorited status.
      operationId: listActivities
      parameters:
        - $ref: '#/components/parameters/Search'
        - in: query
          name: category
          schema:
            $ref: '#/components/schemas/ActivityCategory'
        - in: query
          name: club_id
          schema:
            type: integer
        - in: query
          name: location
          schema:
            type: string
        - in: query
          name: date_from
          schema:
            type: string
            format: date
        - in: query
          name: date_to
          schema:
            type: string
            format: date
        - in: query
          name: only_favorites
          schema:
            type: boolean
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paged list of activities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityListResponse'
    post:
      tags: [Activities]
      summary: Create a new activity
      operationId: createActivity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityCreate'
      responses:
        '201':
          description: Activity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /activities/{activityId}:
    get:
      tags: [Activities]
      summary: Get activity details
      operationId: getActivityById
      parameters:
        - $ref: '#/components/parameters/ActivityId'
      responses:
        '200':
          description: Activity detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Activities]
      summary: Update an activity
      operationId: updateActivity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ActivityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityUpdate'
      responses:
        '200':
          description: Activity updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Activities]
      summary: Delete an activity
      operationId: deleteActivity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ActivityId'
      responses:
        '204':
          description: Activity deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /activities/{activityId}/join:
    post:
      tags: [Registrations]
      summary: Join an activity
      description: Creates or updates the current user's registration for the target activity.
      operationId: joinActivity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ActivityId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinActivityRequest'
      responses:
        '200':
          description: Registration stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: No free slots or registration conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /activities/{activityId}/leave:
    post:
      tags: [Registrations]
      summary: Leave an activity
      description: Marks the current user's registration as left/cancelled.
      operationId: leaveActivity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ActivityId'
      responses:
        '200':
          description: Registration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /activities/{activityId}/registrations:
    get:
      tags: [Registrations]
      summary: List registrations for an activity
      description: Intended for club-side participation monitoring.
      operationId: listActivityRegistrations
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ActivityId'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/RegistrationStatus'
      responses:
        '200':
          description: List of registrations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /activities/{activityId}/favorite:
    put:
      tags: [Favorites]
      summary: Favorite an activity
      operationId: favoriteActivity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ActivityId'
      responses:
        '204':
          description: Activity favorited
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Favorites]
      summary: Remove activity from favorites
      operationId: unfavoriteActivity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ActivityId'
      responses:
        '204':
          description: Favorite removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ActivityId:
      in: path
      name: activityId
      required: true
      schema:
        type: integer
...
    Search:
      in: query
      name: search
      schema:
        type: string
      description: Free-text search term
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: page_size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    User:
      type: object
      required: [user_id, username, email]
      properties:
        user_id:
          type: integer
          example: 12
        username:
          type: string
          maxLength: 50
          example: shannon
        email:
          type: string
          format: email
          maxLength: 100
          example: shannon@example.com
    UserSummary:
      type: object
      required: [user_id, username]
      properties:
        user_id:
          type: integer
        username:
          type: string
          example: shannon
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
          maxLength: 100
        password:
          type: string
          minLength: 8
          format: password
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      required: [access_token, token_type, user]
      properties:
        access_token:
          type: string
          example: eyJhbGciOi...
        token_type:
          type: string
          example: Bearer
        user:
          $ref: '#/components/schemas/User'
    ClubBase:
      type: object
      required: [club_name, club_description]
      properties:
        club_name:
          type: string
          maxLength: 100
          example: FHNW Chess Club
        club_description:
          type: string
          example: Weekly chess meetups and tournaments.
        image_path:
          type: string
          maxLength: 255
          nullable: true
          example: /media/clubs/chess.png
...
    ClubSummary:
      allOf:
        - type: object
          required: [club_id]
          properties:
            club_id:
              type: integer
              example: 3
        - $ref: '#/components/schemas/ClubBase'
        - type: object
          properties:
            is_favorited:
              type: boolean
              example: true
...
    ActivityCategory:
      type: string
      description: Extend this enum as your project categories become final.
      enum:
        - sports
        - club
        - workshop
        - social
        - academic
        - other
    ActivityBase:
      type: object
      required:
        - activity_title
        - activity_description
        - activity_date
        - start_time
        - end_time
        - activity_location
        - activity_category
        - max_slots
        - club_id
      properties:
        activity_title:
          type: string
          maxLength: 100
          example: Volleyball Training
        activity_description:
          type: string
          example: Weekly indoor training session for all skill levels.
        activity_date:
          type: string
          format: date
          example: '2026-05-14'
        start_time:
          type: string
          format: time
          example: '18:00:00'
        end_time:
          type: string
          format: time
          example: '20:00:00'
        activity_location:
          type: string
          maxLength: 100
          example: FHNW Gym Hall A
        max_slots:
          type: integer
          minimum: 1
          example: 24
        activity_category:
          $ref: '#/components/schemas/ActivityCategory'
        image_path:
          type: string
          maxLength: 255
          nullable: true
          example: /media/activities/volleyball.png
        club_id:
          type: integer
          example: 5
    ActivityCreate:
      allOf:
        - $ref: '#/components/schemas/ActivityBase'
    ActivityUpdate:
      type: object
      properties:
        activity_title:
          type: string
          maxLength: 100
        activity_description:
          type: string
        activity_date:
          type: string
          format: date
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        activity_location:
          type: string
          maxLength: 100
        max_slots:
          type: integer
          minimum: 1
        activity_category:
          $ref: '#/components/schemas/ActivityCategory'
        image_path:
          type: string
          maxLength: 255
          nullable: true
        club_id:
          type: integer
    ActivitySummary:
      allOf:
        - type: object
          required: [activity_id]
          properties:
            activity_id:
              type: integer
              example: 11
        - $ref: '#/components/schemas/ActivityBase'
        - type: object
          properties:
            club:
              $ref: '#/components/schemas/ClubSummary'
            joined_count:
              type: integer
              example: 18
            available_slots:
              type: integer
              example: 6
            is_favorited:
              type: boolean
              example: false
            my_registration_status:
              $ref: '#/components/schemas/RegistrationStatus'
    ActivityDetail:
      allOf:
        - $ref: '#/components/schemas/ActivitySummary'
    RegistrationStatus:
      type: string
      enum: [joined, left, cancelled]
      example: joined
    Registration:
      type: object
      required:
        - registration_id
        - registration_status
        - joined_at
        - user_id
        - activity_id
      properties:
        registration_id:
          type: integer
          example: 101
        registration_status:
          $ref: '#/components/schemas/RegistrationStatus'
        joined_at:
          type: string
          format: date-time
          example: '2026-05-10T17:22:00Z'
        user_id:
          type: integer
          example: 12
        activity_id:
          type: integer
          example: 11
        user:
          $ref: '#/components/schemas/UserSummary'
    JoinActivityRequest:
      type: object
      properties:
        status:
          type: string
          enum: [joined]
          default: joined
...
    PaginationMeta:
      type: object
      required: [page, page_size, total_items, total_pages]
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_items:
          type: integer
          example: 58
        total_pages:
          type: integer
          example: 3
    ActivityListResponse:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ActivitySummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    RegistrationListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Registration'
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Resource not found
        code:
          type: string
          example: not_found
        details:
          type: object
          additionalProperties: true
          nullable: true